+++ lib/freecad_stub_gen/generate.py
@@ -7,6 +7,10 @@ from freecad_stub_gen.config import SOURCE_DIR, TARGET_DIR
 from freecad_stub_gen.file_functions import genCppFiles, genXmlFiles
 from freecad_stub_gen.FreeCADTemplates import additionalPath
 from freecad_stub_gen.generators.common.gen_base import BaseGenerator
+from freecad_stub_gen.generators.common.names import (
+    convertNamespaceToModule,
+    getModFromAlias,
+)
 from freecad_stub_gen.generators.exceptions.gen import ExceptionGenerator
 from freecad_stub_gen.generators.from_cpp.functions import (
     FreecadStubGeneratorFromCppFunctions,
@@ -16,7 +20,6 @@ from freecad_stub_gen.generators.from_cpp.module import (
     FreecadStubGeneratorFromCppModule,
 )
 from freecad_stub_gen.generators.from_xml.full import FreecadStubGeneratorFromXML
-from freecad_stub_gen.module_namespace import moduleNamespace
 from freecad_stub_gen.python_code.module_container import Module
 
 logger = logging.getLogger(__name__)
@@ -195,7 +198,7 @@ def improve_Mod(sourcesRoot: Module, sourcePath: Path):
         if moduleName == 'Test':
             continue
 
-        moduleName = moduleNamespace.convertNamespaceToModule(moduleName)
+        moduleName = convertNamespaceToModule(moduleName)
         _genModule(sourcesRoot, mod / 'App', sourcePath, moduleName=moduleName)
         _genModule(sourcesRoot, mod / 'Gui', sourcePath, moduleName=moduleName)
 
@@ -219,7 +222,7 @@ def generateFreeCadStubs(sourcePath=SOURCE_DIR, targetPath=TARGET_DIR):
 
     for stubPackage in targetPath.iterdir():
         if stubPackage.is_dir():
-            if mod := moduleNamespace.getModFromAlias(stubPackage.name):
+            if mod := getModFromAlias(stubPackage.name):
                 newName = mod
             else:
                 newName = stubPackage.name