+++ lib/freecad_stub_gen/generators/common/gen_property/macro/setter_type.py
@@ -8,8 +8,10 @@ from freecad_stub_gen.generators.common.gen_property.macro.alias import (
     PropertyTypeVar,
 )
 from freecad_stub_gen.generators.common.gen_property.macro.base import PropertyMacroBase
-from freecad_stub_gen.generators.common.names import useAliasedModule
-from freecad_stub_gen.module_namespace import moduleNamespace
+from freecad_stub_gen.generators.common.names import (
+    convertNamespaceToModule,
+    useAliasedModule,
+)
 
 logger = logging.getLogger(__name__)
 seen = set[str]()
@@ -174,7 +176,7 @@ class PropertyMacroSetter(PropertyMacroBase):
                 innerType = ''  # read only
 
             case "Mesh::PropertyMeshKernel":
-                mod = moduleNamespace.convertNamespaceToModule('Mesh')
+                mod = convertNamespaceToModule('Mesh')
                 innerType = f'{mod}.Mesh | list[list[float]]'
 
             case "Part::PropertyGeometryList":