+++ lib/freecad_stub_gen/generators/common/return_type_converter/str_wrapper.py
@@ -27,6 +27,9 @@ class AffixCmp:
             case _:
                 raise NotImplementedError
 
+    def __str__(self):
+        return self.text
+
 
 class StrWrapperMeta(type):
     def __call__(cls, *args, **kwargs):