+++ lib/freecad_stub_gen/generators/from_cpp/module.py
@@ -2,8 +2,8 @@ import re
 from collections.abc import Iterable
 
 from freecad_stub_gen.generators.common.cpp_function import findFunctionCall
+from freecad_stub_gen.generators.common.names import convertNamespaceToModule
 from freecad_stub_gen.generators.from_cpp.base import BaseGeneratorFromCpp
-from freecad_stub_gen.module_namespace import moduleNamespace
 from freecad_stub_gen.ordered_set import OrderedStrSet
 from freecad_stub_gen.python_code.module_container import Module
 
@@ -26,7 +26,7 @@ class FreecadStubGeneratorFromCppModule(BaseGeneratorFromCpp):
                 if self._modName is None:
                     raise TypeError
                 curModName = moduleName if '.' in moduleName else self._modName
-                curModName = moduleNamespace.convertNamespaceToModule(curModName)
+                curModName = convertNamespaceToModule(curModName)
 
                 mod[curModName].update(Module(header + result, self.requiredImports))
                 self.requiredImports = OrderedStrSet()